--- monitor/switching.c
@@ -1,2 +1,0 @@
-#include "switching.h"
-
