--- monitor/switching.h
@@ -1,5 +1,0 @@
-#ifndef __SWITCHING_H__
-#define __SWITCHING_H__
-
-
-#endif	// __SWITCHING_H__
